Real-Time-Watcher/requirements.txt
hydralit==1.0.12
hydralit-components==1.0.9
mediapipe==0.8.9.1
numpy==1.21.5
opencv-contrib-python==4.5.5.64
Pillow==9.0.1
streamlit==1.7.0